openai_on_custom_dataset/orchestrator/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
azure-kusto-data==4.0.2
pandas
pyodbc